src/hydra-core/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile setup.cfg
#
asgiref==3.4.1
    # via django
django==3.2.9
    # via
    #   django-redis
    #   djangorestframework
    #   hydra-core (setup.cfg)
django-environ==0.8.1
    # via hydra-core (setup.cfg)
django-redis==5.1.0
    # via hydra-core (setup.cfg)
django-rest-framework==0.1.0
    # via hydra-core (setup.cfg)
djangorestframework==3.13.0
    # via django-rest-framework
paste==3.5.0
    # via hydra-core (setup.cfg)
psycopg2==2.9.2
    # via hydra-core (setup.cfg)
pydantic==1.8.2
    # via hydra-core (setup.cfg)
pytz==2021.3
    # via djangorestframework
redis==3.5.3
    # via django-redis
six==1.16.0
    # via paste
sqlparse==0.4.2
    # via django
typing-extensions==4.0.1
    # via pydantic
waitress==2.0.0
    # via hydra-core (setup.cfg)
whitenoise==5.3.0
    # via hydra-core (setup.cfg)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
